CIJ_sphere: Update for GMT5
Update CIJ_sphere to use GMT5-comptatible syntax.
This is a breaking change from the last commit, and
hence we now require GMT5+.

Specifically, GMT5 introduces a new syntax for
specifying vectors which makes it easier to draw
the ticks for fast orientations onto the sphere
using the `-SJ` option for oriented rectangles.
Replace the use of `-SV` and fix a problem where
GMT draws the ticks which should be invisible
as well as those on the side of the sphere facing
the viewer.